### 2018-06-04 #762 @flor14
Hello.
About the reviewers:
aj2duncan could be a possible reviewer as he mentions experience with rmarkdown and shiny.
Also jhollist interests are related to my field of work.
