### 2018-06-11 #762 @flor14
I had select different names for the files, I have corrected them now. There is a paper.bib and a paper.md
