### 2018-06-19 #762 @karthik
@kylehamilton Tagging you here following my email request. Would you be able to review this Shiny app?
