### 2018-06-19 #762 @kylehamilton
@karthik Yes, I am available and can review this Shiny app.
