### 2018-06-19 #785 @whedon
[REVIEW]: ShinySSD: Species Sensitivity Distributions for Ecotoxicological Risk Assessment
**Submitting author:** @flor14 (<a href="http://orcid.org/0000-0002-0041-097X">María Florencia D&#39;Andrea</a>)
**Repository:** <a href="https://github.com/flor14/paper" target ="_blank">https://github.com/flor14/paper</a>
**Version:** v1.0.0
