### 2018-07-12 #785 @arfon
@flor14 the multiple affiliations need quoting. i.e. change this:

```
affiliation: 1,2
```

to this in your `paper.md`

```
affiliation: "1,2"
```
