### 2018-07-12 #785 @flor14
done!

El jue., 12 jul. 2018 a las 14:53, Arfon Smith (<notifications@github.com>)
escribió:

> @flor14 <https://github.com/flor14> the multiple affiliations need
> quoting. i.e. change this:
>
> affiliation: 1,2
>
> to this in your paper.md
>
> affiliation: "1,2"
>
