### 2018-07-13 #785 @flor14
Great.

Only one thing:
The citation that says "Plan protection Products" is not correct, I have
changed the bib file in order to modify that

Sorry about that and thank you

El jue., 12 jul. 2018 a las 16:25, whedon (<notifications@github.com>)
escribió:

> --> Check article proof 📄 <--
> <https://github.com/openjournals/joss-papers/blob/joss.00785/joss.00785/10.21105.joss.00785.pdf>
>
