### 2018-07-30 #785 @karthik
@kylehamilton Just wanted to check in on the status of your review. Are you waiting on @flor14 for any pending items? 🙏
