### 2018-07-31 #785 @flor14
I have not pending items yet

El lun., 30 jul. 2018 a las 17:13, Karthik Ram (<notifications@github.com>)
escribió:

> @kylehamilton <https://github.com/kylehamilton> Just wanted to check in
> on the status of your review. Are you waiting on @flor14
> <https://github.com/flor14> for any pending items? 🙏
>
