### 2018-08-07 flor14/paper#2 @kylehamilton
Make this into an R package 
This is clearly a novel and useful Shiny app but this would really be a good R package. That would allow things like tests to be run as well as control which packages are loaded into R. For example, the number of conflicts when loading the Shiny app isn't really good form.
![image](https://user-images.githubusercontent.com/2274317/43775806-cf6f15f2-9a02-11e8-9180-b748c8220b74.png)
