### 2018-08-07 flor14/paper#3 @kylehamilton
Tests
As mentioned in #2 once this is made into an R package with the primary parts of the shiny app rolled into functions then we can add coverage tests to make sure everything works smoothly as the R ecosystem changes over time.
