### 2018-08-07 flor14/paper#4 @kylehamilton
Source Code DOI
In the readme you mentioned that the source code has a DOI but it doesn't appear here.
![image](https://user-images.githubusercontent.com/2274317/43776041-852e05f6-9a03-11e8-9a3b-26b46eabe988.png)
