### 2018-08-07 flor14/paper#6 @kylehamilton
Version number needed
There should be something somewhere that lets the user know what version the software is. In an R package this would be in the DESCRIPTION file, with a Shiny app I've also seen this put somewhere in the UI and the README file.
