### 2018-08-07 flor14/paper#7 @kylehamilton
Community guidelines
This needs some community guidelines for people who wish to report issues or submit code to this project.
