### 2018-09-24 #785 @kylehamilton
Sorry @karthik I posted the things that needed to be addressed in the issues of the repo https://github.com/flor14/paper/issues

1. Make the software into an R package so everything would be contained and centralized https://github.com/flor14/paper/issues/2
2. Add tests https://github.com/flor14/paper/issues/3
3. Just want to make sure that there is a DOI for the source code I didn't see this in the .bib file or in the readme https://github.com/flor14/paper/issues/4
4. A minor change would in the name of the repo for production code should be changed from paper to ShinySSD
5. Needs version numbers, shifting this to an R package would help fix this https://github.com/flor14/paper/issues/6
6. I did not see community guidelines https://github.com/flor14/paper/issues/7

If I did this wrong then I'm really sorry I thought that if I posted them as issues in the repo then it would move on from there.
