### 2018-10-24 flor14/shinyssd#3 @flor14
Done.
Some important things:
- For testing I have used shinytest. I divide the test into different little tests. The one that corresponds to /inst/shiny/tests/test_cyper_NOEC_boot.R fails due to changes in the rounding of the numbers.
- test folder is inside inst/shiny
