### 2018-10-24 #785 @flor14
I have completed the review.
I change a little bit the README.md, it is different than the paper.md for publication.

Thank you
