### 2018-11-19 flor14/shinyssd#6 @karthik
Agree with what Kyle says. Please post a version number in the README.
