### 2018-11-19 flor14/shinyssd#7 @karthik
Agreed. This should be easy to do. If you install the `usethis` package, there is a function that can easily template this for you: `usethis::use_tidy_contributing()`
