### 2018-11-19 #785 @karthik
@flor14 I noticed there are couple more issues to fix before we can proceed with this. Can you please address the last two issues (5/6) that Kyle raised. I'd like to wrap this up before the end of the year. Thanks.
