### 2018-11-20 #785 @flor14
I will have a look a soon as possible. Thank you.

El lun., 19 nov. 2018 a las 16:56, Karthik Ram (<notifications@github.com>)
escribió:

> @flor14 <https://github.com/flor14> I noticed there are couple more
> issues to fix before we can proceed with this. Can you please address the
> last two issues (5/6) that Kyle raised. I'd like to wrap this up before the
> end of the year. Thanks.
>
