### 2018-12-12 flor14/shinyssd#6 @flor14
Done. Thank you for all your valuable suggestions. I am happy to leave in GitHub these open source code as part of the results of my PhD and this is because of JOSS existence.
