### 2018-12-19 flor14/shinyssd#6 @flor14
I found a problem in the app, I will fix it soon.
