### 2019-01-25 #785 @flor14
Hello. The last 4 months I have been finishing my PhD so I had no time to
check this (and I will not have until the last week of February). I am not
sure how is JOSS policy about timing, but I can test everything again and
make improvements on February. Sorry about this, it took a bit longer than
expected.
Regards

El sáb., 19 ene. 2019 a las 19:23, Lorena A. Barba (<
notifications@github.com>) escribió:

> 👋 @kylehamilton <https://github.com/kylehamilton> — the author states
> they have made changes addressing the issues you brought up in review.
> Could you have a look, and see if you can check off a few more items from
> the review checklist? If you are satisfied, we'll need a statement from you
> here saying you recommend publication. Thanks! 🙏
>
