### 2019-02-22 #785 @flor14
Well, I finished improving the app, I hope I have interpreted the comments
correctly. Everything seems to work well.
Thanks for the valuable suggestions.

El jue., 14 feb. 2019 a las 7:35, Lorena A. Barba (<notifications@github.com>)
escribió:

> Hi @flor14 <https://github.com/flor14> — Sure. Take your time, but ping
> us back here when you are ready for the review to continue.
>
