### 2019-03-20 #785 @flor14
Hello. Maybe he is not receiving the notifications? Is not possible to send
him an email?

El sáb., 16 mar. 2019 a las 22:07, Lorena A. Barba (<
notifications@github.com>) escribió:

> Hi @kylehamilton <https://github.com/kylehamilton> 👋 — It looks like
> we're waiting for you to check the author's revision and responses to your
> comments. Can you give us a status update?
>
