### 2019-05-20 #785 @karthik
@flor14 Can you please fix the missing DOIs for the references noted above? After that I will proceed with the next steps to getting your paper accepted.
