### 2019-05-20 #785 @flor14
I have just added doi and url to paper.bib. I suppose it solves this

Thank you

Sent from Mail for Windows 10

From: Karthik Ram
Sent: lunes, 20 de mayo de 2019 3:58 p. m.
To: openjournals/joss-reviews
Cc: Florencia; Mention
Subject: Re: [openjournals/joss-reviews] [REVIEW]: ShinySSD: SpeciesSensitivity Distributions for Ecotoxicological Risk Assessment (#785)

@flor14 Can you please fix the missing DOIs for the references noted above? After that I will proceed with the next steps to getting your paper accepted.
