### 2019-05-20 #785 @flor14
I fixed this again according to the autor’s guidelines https://joss.readthedocs.io/en/latest/submitting.html
Now?

Thank you!

Sent from Mail for Windows 10

From: whedon
Sent: lunes, 20 de mayo de 2019 4:34 p. m.
To: openjournals/joss-reviews
Cc: Florencia; Mention
Subject: Re: [openjournals/joss-reviews] [REVIEW]: ShinySSD: SpeciesSensitivity Distributions for Ecotoxicological Risk Assessment (#785)

OK DOIs

- None

MISSING DOIs

- https://doi.org/10.1016/j.ecoenv.2016.08.023 may be missing for title: Aquatic toxicity structure-activity relationships for the zwitterionic surfactant alkyl dimethyl amine oxide to several aquatic species and a resulting species sensitivity distribution
- https://doi.org/10.2903/j.efsa.2013.3290 may be missing for title: Guidance on tiered risk assessment for plant protection products for aquatic organisms in edge-of-field surface waters
- https://doi.org/10.18637/jss.v064.i04 may be missing for title: fitdistrplus: An R Package for Fitting Distributions
- https://doi.org/10.1002/etc.5620190233 may be missing for title: Applying species-sensitivity distributions in ecological risk assessment: assumptions of distribution type and sufficient numbers of species
- https://doi.org/10.1201/9781420032314 may be missing for title: Species sensitivity distributions in ecotoxicology
- https://doi.org/10.1016/s0025-326x(01)00327-7 may be missing for title: Species sensitivity distributions: data and model choice
- https://doi.org/10.1002/9780470057339.vae043.pub2 may be missing for title: EnvStats:  An R Package for Environmental Statistics

INVALID DOIs

- None
