### 2019-05-24 #785 @flor14
After the failed attempts, I regenerated the file again. I added again the
DOIs now to this new file.

El jue., 23 may. 2019 a las 16:46, Karthik Ram (<notifications@github.com>)
escribió:

> @flor14 <https://github.com/flor14> Looks like you removed all the dois?
>
