### 2019-05-28 #785 @flor14
A subindex is missing, also the format of the citation is not correct.
Can I change it or is late?

El mar., 28 may. 2019 a las 15:23, whedon (<notifications@github.com>)
escribió:

> 👉 Check article proof 📄 👈
> <https://github.com/openjournals/joss-papers/blob/joss.00785/joss.00785/10.21105.joss.00785.pdf>
>
