### 2019-05-28 #785 @kyleniemeyer
@flor14 please go ahead and make any changes to the paper
