### 2019-05-29 #785 @flor14
I made a new change. I realize subscripts should be included also as a
formula with $. Hope it works.

El mié., 29 may. 2019 a las 12:53, Kyle Niemeyer (<notifications@github.com>)
escribió:

> @flor14 <https://github.com/flor14> thanks! I will make the final checks
> on your submission later today.
>
