### 2019-05-29 #785 @kyleniemeyer
Hi @flor14, could you add full affiliation information as in a typical article (e.g., department, city, country)?

Also, I made a small PR to fix a few things, if you could accept and merge that: https://github.com/flor14/shinyssd/pull/9
