### 2019-05-29 #785 @flor14
I improved what you mentioned. If citations are ok, and subscripts are
correct in the second paragraph: " HC1, HC5 and HC10"  for me it is ready.
Maybe citations could be more detailed in the author's guidelines?
Mentioning the 3 main cases:
1- @author:2001  ->   author et al. (2001)
2- [@author:2001] -> (author et al. 2001)
3- [@author1:2001; @author2:2001]  -> (author1 et al. 2001; author2 et al.
2002)?
I mention this just to save your time.

Thank you Kyle! You was very kind

El mié., 29 may. 2019 a las 16:40, Kyle Niemeyer (<notifications@github.com>)
escribió:

> Hi @flor14 <https://github.com/flor14>, could you add full affiliation
> information as in a typical article (e.g., department, city, country)?
>
> Also, I made a small PR to fix a few things, if you could accept and merge
> that: flor14/shinyssd#9 <https://github.com/flor14/shinyssd/pull/9>
>
