### 2019-05-29 #785 @kyleniemeyer
@flor14 yes, the citation documentation could be improved... I've had to fix this issue in a number of papers.
