### 2019-05-29 #785 @kyleniemeyer
@flor14 for the last subscript, I think you need to have `HC_{10}` (or however you need it styled)
