### 2019-05-29 #785 @flor14
I've changed subscripts to this now:  $HC_1$, $HC_5$ and $HC_{10}$
The problem seems the last one with 2 digits

El mié., 29 may. 2019 a las 18:29, whedon (<notifications@github.com>)
escribió:

> PDF failed to compile for issue #785
> <https://github.com/openjournals/joss-reviews/issues/785> with the
> following error:
>
> Error producing PDF.
> ! Double subscript.
> l.316 ...estimators (HC_1), (HC_5) and (HC_1_
>
> Looks like we failed to compile the PDF
>
