### 2019-05-29 #785 @kyleniemeyer
Congrats @flor14  on your submission's publication in JOSS! Thanks to @kylehamilton for reviewing and @karthik for editing!
